Backend/requirements.txt
altgraph==0.17.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-auth-ldap==4.1.0
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-debug-toolbar-mongo==0.1.10
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.7
Faker==19.2.0
inflection==0.5.1
packaging==23.1
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ldap==3.4.3
pytz==2022.2.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.7.1
uritemplate==4.1.1
